# server/docs/swagger.yaml
basePath: /
definitions:
...
  api.torznabCapsCategory:
    properties:
      id:
        type: string
      name:
        type: string
      subcategories:
        items:
          $ref: '#/definitions/api.torznabCapsCategory'
        type: array
    type: object
  api.torznabCapsResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/api.torznabCapsCategory'
        type: array
      limits:
        properties:
          default:
            type: integer
          max:
            type: integer
        type: object
    type: object
...
  models.TorrentDetails:
    properties:
      audioQuality:
        type: integer
      categories:
        type: string
      createDate:
        type: string
      hash:
        type: string
      imdbid:
        type: string
      link:
        type: string
      magnet:
        type: string
      name:
        type: string
      names:
        items:
          type: string
        type: array
      peer:
        type: integer
      seed:
        type: integer
      size:
        type: string
      title:
        type: string
      tracker:
        type: string
      videoQuality:
        type: integer
      year:
        type: integer
    type: object
...
info:
  contact: {}
  description: Torrent streaming server.
  license:
    name: GPL 3.0
  title: Swagger Torrserver API
  version: '{version.Version}'
paths:
...
  /torznab/caps:
    get:
      description: Returns the categories and limits advertised by a configured indexer.
      parameters:
      - description: Indexer index in the configured list
        in: query
        name: index
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Indexer capabilities
          schema:
            $ref: '#/definitions/api.torznabCapsResponse'
      summary: Fetch torznab indexer capabilities
      tags:
      - API
  /torznab/search:
    get:
      description: Makes a torznab search.
      parameters:
      - description: Torznab query
        in: query
        name: query
        required: true
        type: string
      - description: Indexer index (-1 = all configured)
        in: query
        name: index
        type: string
      - description: Torznab categories, comma-separated (default 5000,2000 = TV+Movies)
        in: query
        name: cat
        type: string
      - description: Result offset for pagination
        in: query
        name: offset
        type: string
      - description: Max results (default 100, capped at 250)
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Torznab torrent search result(s)
          schema:
            items:
              $ref: '#/definitions/models.TorrentDetails'
            type: array
      summary: Makes a torznab search
      tags:
      - API
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"